main: add retry for server if it returns
If the server dies, e.g. the interface it is bound to goes away, then
this will retry starting it. The interval of 1 second might be a bit too
aggressive, but it allows for it to be fairly responsive if a cable is
replugged. This could be configured with a cmdline option in a future
patch.