tests: add simple fuzzer for dhcp_get_request_type
This stuffs random data into the options field in the request, and then
calls dhcp_get_request_type in a loop. Should be run in gdb so that any
exceptions are caught.

Originally I had looked at the return value and if it wasn't -1 then
throw an error, however I found that dhcp_get_request_type could find a
valid DHCP REQUEST/DISCOVER byte sequence fairly often (it's only 3
bytes... so not complicated...)