rename header->message, use correct option length
I should have used 'message' all along, since that's what the relevant
RFC calls this. The 'header' is the first 236 bytes before the
'options'.

The min/max size for the message are added as macros, and the length of
the options was adjusted based on further reading of RFC2131.

Lastly, this fixes an out-of-bounds access that was not caught by
-Wall -Wextra, but only showed up with -Os